feat: Render html on behalf of the initial browser
When a crawling-job runs automatically, we need to re-render the page (to which it refers to) on behalf of the initial user's browser, and after the javascript has run. This is required as there are cases where certain HTML elements, and their css, are rendered differently in different screens. To tackle that 1) we store the width, the height and the agent of the user's browser in the CrawlingJob model, and 2) we use the playwright module that fetches tha page with all its javascript loaded, using these three settings.